requirements/base.txt
Django==1.8.2
django-formtools==1.0
djangorestframework==3.1.3
django-filter==0.10.0
django-widget-tweaks==1.3
Markdown==2.6.2
phonenumbers==7.0.4
Pygments==2.0.2
pytz==2015.4
requests==2.7.0
threadfix-api==1.1.1
